+++ objectbox/example/flutter/objectbox_demo_sync/lib/objectbox.dart
@@ -42,4 +42,23 @@ class ObjectBox {
         );
     return ObjectBox._create(store);
   }
+
+  /// Add a note within a transaction.
+  ///
+  /// To avoid frame drops, run ObjectBox operations that take longer than a
+  /// few milliseconds, e.g. putting many objects, in an isolate with its
+  /// own Store instance.
+  /// For this example only a single object is put which would also be fine if
+  /// done here directly.
+  Future<void> addNote(String text) =>
+      store.runInTransactionAsync(TxMode.write, _addNoteInTx, text);
+
+  /// Note: due to [dart-lang/sdk#36983](https://github.com/dart-lang/sdk/issues/36983)
+  /// not using a closure as it may capture more objects than expected.
+  /// These might not be send-able to an isolate. See Store.runAsync for details.
+  static void _addNoteInTx(Store store, String text) {
+    // Perform ObjectBox operations that take longer than a few milliseconds
+    // here. To keep it simple, this example just puts a single object.
+    store.box<Note>().put(Note(text));
+  }
 }